Preserving form state in Find Probes
The state of YUI radio buttons is not saved when the user hits
'Back' browser button to go back from results to the form page.
This is solved by adding a hidden field which remembers the index
of the button selected, and then, on window load event, the correct
button is reselected from the saved index.