UploadGO working!
There may be a bug that has to do with how records are counted and
displayed in the user message (though it does not seem to affect
actual results) -- the number of GO terms updated may be reported
as double of what it actually was. Will work on this later.