Uploading probe worked out; Wait indicators...
Semi-final upload probe list code in ManagePlatforms -- still need
to make the sequence column optional and to test whether the update
version (not create) is working fine.

Deleting platforms may take a while, so showing a loading indicator
as a panel with an AJAX request gif image from the YUI library.