Settled on upload interface in ManageSpecies
This interface will also be used in ManagePlatforms for uploading
platform annotation. The idea is that the user has a row of buttons
representing columns in a file to choose from. Below the row of
buttons, a banner box displays which columns are to be present in the
uploaded file. The banner box is updated depending on which buttons
are selected.