Streamlined CompExp
Moved a lot of JS code away from the Compare Experiments Perl module
into a standalone file. URI to Google API Venn diagrams are now built
using Javascript instead of Perl. Reusing the same data structure that
is used to populate DataTable in main form in Compare Experiments for
the summary table on the subsequent page.