USe hidden field instead of SELECT in species filter in CompExp
Because species is already preselected whenever a Platform is chosen in
the top drop down in Compare Experiments form, it is no longer
necessary to let user use species drop down in query filter, or even to
display it at all. Therefore, if species data are missing among
arguments to mainFormDD, do not show the species drop down at all and
use a hidden field instead.