Platform/Study drop downs now reflect current project
Previously, in Compare Experiments and OutputData, drop downs for
platform and study would show all platforms and all studies regardless
of what the current project is set to. Now limiting options in the drop
downs to those studies that actually belong to the current project and
to those platform that have at least one study that belongs to the
current project.