Bugfix in TFSDisplay / CSV report
Implemented correct procedure for expected value calculation for rows
in the breakdown table in CSV reports: extract the value of TFS before
the dot (e.g. if TFS is 7.121, then the value extracted is 7), count
the number of bits in it (which is the number of significant
experiments), find the number of permutations by raising 2 to the power
of the number of bits (1 << 3 == 8 in this case), and divide the
expected number for the group (for all 7s) by the number of
permutations possible.