Fix in OutputData and some refactoring
In Output data, now making two queries: one to get experiment info, and
another to get data itself. The results are merged on output but not
stored in memory as merged structure. Elsewhere (FindProbes,
CompareExperiments, TFSDisplay) as well as in OutputData, moved the
routine that finds out current project information to DBLists; renamed
DBLists itself to DBHelper.