One bugfix plus small changes
Fixed: In Compare Experiments, when the same experiment was compared to
itself using two different P-values, previously would get an error
message. This is no longer so. Added: added a P-value label (simply a
number at this point) to the experiment heading in the TFS breakdown
table. Added paragraph to the About page. Added more text to the Sign
Up page. Removed: will not show plain flag sum (with no decimal part)
in CSV report.