Changes to Session::User, SGX::Profile
The biggest improvement: user is not created at the registration time
anymore. Instead, the user is set up during the first email
confirmation cycle. This now means that we cannot have self-created
users without (confirmed) emails (unless they were created by the
administrator). This also means that the mail server (Sendmail,
Postfix) must always be up for new users to be able to register
themselves. TODO: the very same behavior should be expanded to
verifyEmail; to be done in the next commit.