Introduced SGX::Exception::HTTP
This exception if exclusively for HTTP errors and has a custom field
"status" which can be set to values like '404' for example. It is
handled by sending the corresponding status to the browser window and
showing error page body through SGX::Static. SGX::Config is now an
object (instead of an imported module).