Do not raise error if same experiment/pval combo is compared
TFSDisplay would previously throw an error if same experiment/pval
combo occurred more than once in the comparison. Such comparisons are
nonsensical, but there is no sense not to allow them, or, if they are
disallowed, an explicit message should explain the reason why program
cannot continue, instead of doing what it was doing before (a DBI error
would be raised saying that temporary table names conflict with each
other).